Ændret int pris til double pris

+++ src/billetautomat/BenytBilletautomat.java
@@ -8,7 +8,7 @@ public class BenytBilletautomat
 		System.out.println("Installatøren har installeret en ny billetautomat");
 		System.out.println();
 		System.out.println("En kunde vil købe en billet");
-		int pris = automat.getBilletpris();
+		double pris = automat.getBilletpris();
 		System.out.println("Billetprisen er: " + pris + " kroner");
 		System.out.println("Kunden putter 24 kroner i automaten");
 		automat.indsætPenge(24);
